requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
sqlmodel
python-dotenv
passlib[bcrypt]
python-multipart
pyjwt